Removing ssl-hello-chk from kibana backend
The kibana role is laying down an apache configuration that disables v2
and v3 SSL.  The haproxy config for the kibana-ssl backend is trying to
do a SSLv3 hello check which is causing chattines in the apache log
files. This change removes the 'option ssl-hello-chk' line and replaces
it with a 'check' declaration. The playbook is installing haproxy-1.4.24
out of the main trusty repositories, and that version does not support
any the check-ssl functionality.

Connected https://github.com/rcbops/rpc-openstack/issues/1141